role and permission

+++ app/Http/Controllers/TenantsController.php
@@ -15,8 +15,14 @@ class TenantsController extends Controller
 
     public function index()
     {
+        $user = auth()->user();
+
+       
+        if ($user->can('get all tenants')) {
         $tenant = Tenant::all();
         return $this->ApiResponse($tenant , 'get all tenants successfully' , 201 );
+        }else {
+            return $this->ApiResponse( null , 'you do not have permission  ' , 400 ); } 
         
     }
 
@@ -24,41 +30,70 @@ class TenantsController extends Controller
 
     public function store(TenantRequest $request)
     {
-
-        $tenant = Tenant::create($request->all());
+        $user = auth()->user();
 
         
-        return $this->ApiResponse($tenant , 'store tenants successfully' , 201 );
+        if ($user->can('create tenant')) {
+            $tenant = Tenant::create($request->all());
+
+        
+            return $this->ApiResponse($tenant , 'store tenant successfully' , 201 );  
+        }
+        
+        else {
+            return $this->ApiResponse( null , 'you do not have permission  ' , 400 ); } 
 
     }
 
 
     public function show($id)
     {
+        $user = auth()->user();
+
+        
+        if ($user->can('show tenant')) {
         $tenant = Tenant::find($id);
         return $this->ApiResponse($tenant , 'get tenants successfully' , 201 );
+        } else {
+            return $this->ApiResponse( null , 'you do not have permission  ' , 400 ); } 
     }
 
 
     public function update(TenantRequest $request, $id)
     {
+        $user = auth()->user();
+
+        
+        if ($user->can('update tenant')) {
         $tenant = Tenant::find($id);
         $tenant->update($request->all());
         return $this->ApiResponse($tenant , 'updated tenants successfully' , 201 );
+        } else {
+            return $this->ApiResponse( null , 'you do not have permission  ' , 400 ); } 
 
     }
 
 
     public function destroy($id)
     {
+        $user = auth()->user();
+
+        
+        if ($user->can('destroy tenant')) {
         Tenant::destroy($id);
         return $this->ApiResponse( null , 'deleted tenants successfully' , 201 );
+        } else {
+            return $this->ApiResponse( null , 'you do not have permission  ' , 400 ); } 
 
     }
 
     public function assigne(Request $request){
+        $user = auth()->user();
+        if ($user->can('assigne user to tenant')) {
         $user = User::find($request->user_id);
         $user->update(['tenant_id' => $request->tenant_id]);
         return $this->ApiResponse( $user , 'update tenant user successfully' , 201 );
+    } else {
+        return $this->ApiResponse( null , 'you do not have permission  ' , 400 ); } 
     }
 }
